Fixed logic in user_patch and time_entry_patch, updated translation files, fixed relation in FreshbooksProject, fixed timelog edit form hook
Fixed the logic in user_patch to handle cases where user is updated by the admin.  The api_key field is not present on the admin user interface, so the api_key's value is potentially nil.
Updated the time_entry_patch to set the field freshbooks_project_permitted to true if the user has permission in Freshbooks to log hours to the project
Updated the timelog edit form hook to disable the 'Send to Freshbooks' checkbox if the user's freshbooks_project_permitted is false
Added the english translations of the message to be displayed when the timelog edit from 'Send to Freshbooks' checkbox is disabled